requirements.txt
absl-py==0.9.0
altair==4.1.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.5
blinker==1.4
blis==0.4.1
bokeh==2.0.2
boto3==1.13.24
botocore==1.16.24
Bottleneck==1.3.2
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-completion==0.5.1
cloudpickle==1.4.1
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.1
decorator==4.4.2
defusedxml==0.6.0
distributed==2.15.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fastprogress==0.2.3
filelock==3.0.12
fsspec==0.7.1
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
HeapDict==1.0.1
hestia==0.3.1
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
jedi==0.17.0
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.33.0
locket==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.0.0rc5
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
olefile==0.46
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
polystores==0.2.5
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.4
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pydeck==0.4.0b1
Pygments==2.6.1
# pynvx==1.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
raven==6.7.0
requests==2.23.0
requests-toolbelt==0.9.1
rhea==0.5.5
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
shellingham==1.3.2
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0.1
spacy==2.2.4
srsly==1.0.2
streamlit==0.61.0
tabulate==0.8.2
tblib==1.6.0
tensorflow==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
toml==0.10.1
toolz==0.10.0
tornado==5.0
tqdm==4.46.0
traitlets==4.3.3
typing-extensions==3.7.4.1
tzlocal==2.1
Unipath==1.1
validators==0.15.0
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
wrapt==1.12.1
xlrd==1.2.0
zict==2.0.0
zipp==3.1.0
